# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "lessonlink-backend バックエンドAPI",
        "title": "lessonlink-backend",
        "contact": {},
        "license": {
            "name": "typedef-tokyo."
        },
        "version": "1.0"
    },
    "host": "localhost:3002",
    "basePath": "/",
    "paths": {
        "/campus/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "キャンパスリスト取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.CampusListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/lesson/{campus}": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "講座追加",
                "parameters": [
                    {
                        "type": "string",
                        "description": "校舎",
                        "name": "campus",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "講座追加リクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.LessonAddRequestData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.LessonAddResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/lesson/{campus}/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "講座一覧取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "校舎",
                        "name": "campus",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.LessonListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/lesson/{lessonid}": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "summary": "講座編集",
                "parameters": [
                    {
                        "type": "string",
                        "description": "講座",
                        "name": "lessonid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "講座編集リクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.LessonEditRequestData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.LessonEditResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/schedule/{schedule_id}/room/invisible": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "summary": "非表示ルーム登録",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ScheduleID",
                        "name": "schedule_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "非表示ルームリクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.InvisibleRoomSaveRequestData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.InvisibleRoomSaveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/user/self": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "ログインユーザー取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.UserGetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.InvisibleRoomSaveRequestData": {
            "type": "object",
            "required": [
                "invisible_rooms"
            ],
            "properties": {
                "invisible_rooms": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "controller.LessonAddRequestData": {
            "type": "object",
            "required": [
                "duration",
                "lesson_name"
            ],
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "lesson_name": {
                    "type": "string"
                }
            }
        },
        "controller.LessonEditRequestData": {
            "type": "object",
            "required": [
                "duration",
                "lesson_name"
            ],
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "lesson_name": {
                    "type": "string"
                }
            }
        },
...
        "presenter.CampusListDTO": {
            "type": "object",
            "required": [
                "campus",
                "campus_name"
            ],
            "properties": {
                "campus": {
                    "type": "string"
                },
                "campus_name": {
                    "type": "string"
                }
            }
        },
        "presenter.CampusListResponse": {
            "type": "object",
            "required": [
                "campuses",
                "msg"
            ],
            "properties": {
                "campuses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.CampusListDTO"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "presenter.InvisibleRoomSaveResponse": {
            "type": "object",
            "required": [
                "msg"
            ],
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "presenter.LessonAddResponse": {
            "type": "object",
            "required": [
                "msg"
            ],
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "presenter.LessonEditResponse": {
            "type": "object",
            "required": [
                "msg"
            ],
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "presenter.LessonListDTO": {
            "type": "object",
            "required": [
                "id",
                "lesson_duration",
                "lesson_name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "lesson_duration": {
                    "type": "integer"
                },
                "lesson_name": {
                    "type": "string"
                }
            }
        },
        "presenter.LessonListResponse": {
            "type": "object",
            "required": [
                "lessons"
            ],
            "properties": {
                "lessons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.LessonListDTO"
                    }
                }
            }
        },
...
        "presenter.UserGetResponse": {
            "type": "object",
            "required": [
                "id",
                "name",
                "role_key",
                "user_name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "role_key": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
...
    }
}